fix clang W4 warnings.

+++ PKRenderer/src/Core/RHI/Structs.h
@@ -358,9 +358,9 @@ namespace PK
     static constexpr Type operator | (const Type& a, const Type& b) { return (Type)((uint32_t)a | (uint32_t)b); } \
     static constexpr Type operator |= (const Type& a, const Type& b) { return a | b; } \
     static constexpr Type operator & (const Type& a, const Type& b) { return (Type)((uint32_t)a & (uint32_t)b); } \
-    static constexpr Type operator & (const Type& a, const int& b) { return (Type)((uint32_t)a & (uint32_t)b); } \
-    static constexpr bool operator == (const Type& a, const int& b) { return (uint32_t)a == b; } \
-    static constexpr bool operator != (const Type& a, const int& b) { return (uint32_t)a != b; } \
+    static constexpr Type operator & (const Type& a, const uint32_t& b) { return (Type)((uint32_t)a & b); } \
+    static constexpr bool operator == (const Type& a, const uint32_t& b) { return (uint32_t)a == b; } \
+    static constexpr bool operator != (const Type& a, const uint32_t& b) { return (uint32_t)a != b; } \
 
     PK_DECLARE_ENUM_OPERATORS(ColorMask)
     PK_DECLARE_ENUM_OPERATORS(BufferUsage)
